requirements.txt
google-generativeai==0.2.0  # Substitua pela versão correta
pandas==1.5.3
numpy==1.24.0